pyproject.toml
[tool.poetry]
name = "opencv-haar-train"
version = "0.1.0"
description = ""
authors = ["astra <zorkin.roman@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
opencv-contrib-python = "^4.5.5"
Pillow = "^9.1.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
